sentiment-analyzer/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
textblob==0.17.1
nltk==3.8.1
text2emotion==0.0.5
vaderSentiment==3.3.2
python-dotenv==1.0.0
pandas==2.0.3
numpy==1.24.3